feature: create accountTypes component with styles
