+++ src/features/profile/ui/settings/accountManagement/accountTypes/accountTypes.module.scss
@@ -1,0 +1,14 @@
+.account {
+  order: 1;
+  width: 121.09%;
+}
+
+.container {
+  padding: 14px 20px;
+  background-color: var(--dark-500);
+  outline: 1px solid var(--dark-300);
+}
+
+.title {
+  margin-bottom: 5px;
+}
