+++ src/features/profile/ui/settings/accountManagement/accountTypes/index.ts
@@ -1,0 +1,1 @@
+export * from './accountTypes'
